requirements.txt
Flask
plexapi
python-dotenv

